Moved {json_types => types}/public_key.rs
