+++ near-sdk/compilation_tests/bad_argument.stderr
@@ -1,13 +1,3 @@
-warning: use of deprecated type alias `json_types::public_key::Base58PublicKey`: PublicKey type is now unified with Base58PublicKey. It is recommended to use PublicKey going forward to avoid using similar sounding types for the same thing.
-  --> $DIR/mod.rs:12:22
-   |
-12 | pub use public_key::{Base58PublicKey, CurveType, PublicKey};
-   |                      ^^^^^^^^^^^^^^^
-   |
-   = note: `#[warn(deprecated)]` on by default
-
-warning: 1 warning emitted
-
 error: Unsupported argument type.
   --> $DIR/bad_argument.rs:28:1
    |
