+++ near-sdk/compilation_tests/payable_view.stderr
@@ -1,13 +1,3 @@
-warning: use of deprecated type alias `json_types::public_key::Base58PublicKey`: PublicKey type is now unified with Base58PublicKey. It is recommended to use PublicKey going forward to avoid using similar sounding types for the same thing.
-  --> $DIR/mod.rs:12:22
-   |
-12 | pub use public_key::{Base58PublicKey, CurveType, PublicKey};
-   |                      ^^^^^^^^^^^^^^^
-   |
-   = note: `#[warn(deprecated)]` on by default
-
-warning: 1 warning emitted
-
 error: Payable method must be mutable (not view)
   --> $DIR/payable_view.rs:12:5
    |
