+++ near-sdk/src/types/vm_types.rs
@@ -2,7 +2,7 @@
 pub use near_vm_logic::types::{PromiseResult as VmPromiseResult, ReturnData};
 
 //* Types from near_vm_logic
-pub use crate::json_types::PublicKey;
+// pub use super::public_key::PublicKey;
 pub type PromiseIndex = u64;
 pub type ReceiptIndex = u64;
 pub type IteratorIndex = u64;
